Plugins: port JSON plugin
